Add function to report internal client errors
This is to allow gss_display_status to return a meaningful error if
there is an internal client error as opposed to a returned proxy server
error. If we do not do this a call to gpm_display_status() after a failure
would return a NULL string and give no clue about the error.