Local vs Remote cred check fixes
Do not check for remote cred in LOCAL_ONLY case, since we are instructed to
specifically not check with the proxy.  However, still fall back to remote
with cred_store, since it may simply indicate what ccache to use to store the
remote credentials reference and not where to fetch the credential from.

Related-to: #51
Signed-off-by: Simo Sorce <simo@redhat.com>
[rharwood@redhat.com: Shrink commit message and typos]
Reviewed-by: Robbie Harwood <rharwood@redhat.com>