[client] Switch to non-blocking sockets
Switch the gssproxy client library to non-blocking sockets, allowing
for timeout and retry operations.  The client will automatically retry
both send() and recv() operations three times on ETIMEDOUT.  If the
combined send() and recv() hit the three time limit, ETIMEDOUT will be
exposed to the caller in the minor status.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>
Reviewed-by: Simo Sorce <simo@redhat.com>
[rharwood@redhat.com: commit message cleanups, rebased]
Reviewed-by: Robbie Harwood <rharwood@redhat.com>