Always choose highest requested debug level
Allowing the CLI to lower the debug level specified in a config file
is dubious, and previously broken since we don't distinguish "default
value" from "explicitly requested value of 0" in popt.  This resulted
in "Debug Enabled (level: 0)" even when the log level was not actually
0, which is confusing for users.

Remove the gp_debug_args() function since it is no longer used.

Signed-off-by: Robbie Harwood <rharwood@redhat.com>
Reviewed-by: Simo Sorce <simo@redhat.com>
Merges: #229