examples don't play well together
The 80-httpd and 99-network-fs-clients examples might be used together
such as an Apache httpd web server that authenticates web clients but is
also authenticated itself as an NFS client to access remote content it
must serve to web clients.

This prevents ticket collisions in the credential caches by making them
distinct for these two independent use cases so that they might be used
together.

Signed-off-by: John Florian <jflorian@doubledog.org>