Add special user-proxy mode
This mode is to be used in a user session where what we want to proxy
is other applications run by the same user in a container or flatpak.
This allows those application to use the user's creedntials w/o having
to bind mount and expose the user TGT.
The configuration is fixed and simplified, and the user proxy is allowed
to be itself intercepted by the system proxy if desired.

Use a subdirectory of XDG_RUNTIME_DIR by default
If an explicit socket name is passed we expect the directory structure
to be in place.

However if we are dealing with XDG_RUNTIME_DIR then we want to create a
subdirectory so that it can bind mounted in containers.
Bind mounting the socket directly is problematic if we want to allow to
restart the daemon as the socket may then become permanently
disconnected in the container.

If we use a subdirectory we need to attempt to create it on start
because XDG_RUNTIME_DIR is ephemeral and generally recreated from
scratch when the user session is created, so we need to populate it with
the needed dir strcuture ourself to make things simpler for users.

Signed-off-by: Simo Sorce <simo@redhat.com>