Fix issues with socket activation.
Adjust systemd's service file for usermode.
Socket activation requires that we used the interactive mode (-i)
so that gssproxy does not fork on startup.
Forking causes sd_listen_fds() to fail as that function checks that
the pid of the process calling it is the same process systemd forked
via the internal envvar LISTEN_PID.

Additionally make sure to set the socket name for the artificial
config service in usermode as well as the socket set up during socket
activation to all use the same string (the usermode socket) so that
the checks in gp_creds_match_conn() work properly in the socket
activation case too.

Signed-off-by: Simo Sorce <simo@redhat.com>