Fix self termination and add some accounting
Self termination need to happen only when gssproxy is idle for a certina
amount of time.

However due to non material activity the gssproxy performs through the
verto event handler we need to actively measure ideling rather than
depend on things like verto_idle.

Ad an accounting function to our socket handler.
This now allows us to collect some interesting statistics as well as
reset the idle handler when there is actual activity ongoing.

We reset the handler only upon receiving data from a client, as that
indicates a requested activity. Writes happen only as reposnses and are
considered part of a single activity event together with the read that
caused the idel handler reset.

Signed-off-by: Simo Sorce <simo@redhat.com>