Add a wait to redirect Krb5 tracing to our stderr
The proccess stderr may be redirected to a socket or a pipe by the
parent process (which is what systemd does for example).

When that happen we can't simply pass a valid path to stderr so
that libkrb5 can send the tracing via the KRB5_TRACE environment
variable to a our stderr, passing /dev/stderr in this case will
just result in an open error and tracing output going nowhere.

To handle this we create our own fifo to pass to libkrb5 and
then we create a thread that reads the incoming data and pipes
it back to the stderr we've been given by our parent.

Signed-off-by: Simo Sorce <simo@redhat.com>