Move some configure checks out of a condition
I was too aggressive in moving all systemd checks, some of them need to
be performed at all times because there are conditionals ependent on the
detection.

The actual build for gss-only module is controlled by the BUILD_PROXY
guard anyway, so the checking unconditionally should not cause any
change for that configuration even when unused.

Signed-off-by: Simo Sorce <simo@redhat.com>